requirements.txt
altgraph==0.17.4
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.32.3
setuptools==71.1.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
SQLAlchemy==2.0.1
